Now able to (hopefully) successfully send the base64 to the index.php from the base64Handler.js.
So in more detail, now can:
. Drag-drop an image into the box.
. This image will be converted into its base64 - and will also inform
user that it has been successfully converted (by showing some text).
. Once you press the 'Send base64' button, base64Handler.js will print
the base64 to the console and send the base64 to the index.php.
So far, this process of sending it to the index.php has caused the
word "Success" to be printed to the console, which seems good. But will
need to further check whether the base64 is really successfully sent or
not.

. The base64Handler.js is a handler for retrieving the base64 and
sending it to the index.php as a POST type.

. No change of the imageProcessing.py.